+++ lib/CountyPlots.js
@@ -4,6 +4,7 @@ import { useMemo, useCallback } from 'react'
 import { useFetch } from './queries.js'
 import us from 'us'
 import WindowSelectSearch from './WindowSelectSearch.js'
+import { makeUnitsSelect } from '../lib/selects.js'
 
 const fuseOptions = {
   keys: ['name'],
@@ -77,9 +78,11 @@ export default function CountyPlots ({ countyName, stateAbbr, stateCode }) {
     [countyName, stateAbbr, status]
   )
 
+  const { selectedUnits, unitsSelect } = makeUnitsSelect()
+
   return (
     <div className='mx-auto mb-10 align-center items-center flex flex-col justify-center'>
-      <div className='lg:grid lg:grid-cols-5 flex flex-col'>
+      <div className='lg:grid lg:grid-cols-3 flex flex-col'>
         <div className='m-4 col-span-1'>
           <WindowSelectSearch
             search
@@ -89,13 +92,14 @@ export default function CountyPlots ({ countyName, stateAbbr, stateCode }) {
             value={optionVal?.value}
           />
         </div>
-        <div className='mt-4 mb-1 col-span-3 text-center'>
+        <div className='mt-4 mb-1 col-span-1 text-center'>
           <h1 className='text-4xl'>{countyName}, {stateAbbr}</h1>
         </div>
+        <div className='col-span-1 m-4'>{unitsSelect}</div>
       </div>
 
       <div className='w-full flex flex-row'>
-        <BarPlot data={{ table: data }} units='units' perCapita={false} />
+        <BarPlot data={{ table: data }} units={selectedUnits} perCapita={false} />
       </div>
     </div>
   )
