+++ lib/MetroPlots.js
@@ -3,6 +3,7 @@ import BarPlot from './BarPlot.js'
 import { useMemo, useEffect, useCallback } from 'react'
 import { useFetch } from './queries.js'
 import WindowSelectSearch from './WindowSelectSearch.js'
+import { makeUnitsSelect } from '../lib/selects.js'
 
 function getJsonUrl (metro) {
   if (metro === null || (typeof metro === 'undefined')) {
@@ -121,10 +122,11 @@ export default function MetroPlots ({ metroPath, setTitle }) {
   )
   /* eslint-enable */
 
-  // TODO: Add buildings/property option
+  const { selectedUnits, unitsSelect } = makeUnitsSelect()
+
   return (
     <div className='mx-auto mb-10 align-center items-center flex flex-col justify-center'>
-      <div className='lg:grid lg:grid-cols-5 flex flex-col'>
+      <div className='lg:grid lg:grid-cols-3 flex flex-col'>
         <div className='m-4 col-span-1'>
           <WindowSelectSearch
             search
@@ -135,13 +137,14 @@ export default function MetroPlots ({ metroPath, setTitle }) {
             renderOption={renderOption}
           />
         </div>
-        <div className='mt-4 mb-1 col-span-3 text-center'>
+        <div className='mt-4 mb-1 col-span-1 text-center'>
           <h1 className='text-4xl'>{optionVal?.metro_name ?? ''}</h1>
         </div>
+        <div className='col-span-1 m-4'>{unitsSelect}</div>
       </div>
 
       <div className='w-full flex flex-row'>
-        <BarPlot data={{ table: data }} units='units' perCapita={false} />
+        <BarPlot data={{ table: data }} units={selectedUnits} perCapita={false} />
       </div>
       {countyList}
     </div>
