+++ lib/PlacePlots.js
@@ -4,6 +4,7 @@ import WindowSelectSearch from '../lib/WindowSelectSearch.js'
 import BarPlot from './BarPlot.js'
 import us from 'us'
 import { useRouter } from 'next/router'
+import { makeUnitsSelect } from '../lib/selects.js'
 
 function getJsonUrl (place, state) {
   if (place === null) {
@@ -89,9 +90,11 @@ export default function PlacePlots ({ place, state }) {
 
   const isCounty = place ? place.endsWith('County') || place.endsWith('Parish') : false
 
+  const { selectedUnits, unitsSelect } = makeUnitsSelect()
+
   return (
     <div className='mx-auto mb-10 align-center items-center flex flex-col justify-center'>
-      <div className='lg:grid lg:grid-cols-5 flex flex-col'>
+      <div className='lg:grid lg:grid-cols-3 flex flex-col'>
         <div className='m-4 col-span-1'>
           <WindowSelectSearch
             search
@@ -101,14 +104,15 @@ export default function PlacePlots ({ place, state }) {
             value={optionVal}
           />
         </div>
-        <div className='mt-4 mb-1 col-span-3 text-center'>
+        <div className='mt-4 mb-1 col-span-1 text-center'>
           {isCounty && <h2 className='text-2xl -mb-2'>Unincorporated</h2>}
           <h1 className='text-4xl'>{place}, {state}</h1>
         </div>
+        <div className='col-span-1 m-4'>{unitsSelect}</div>
       </div>
 
       <div className='w-full flex flex-row'>
-        <BarPlot data={{ table: data }} units='units' perCapita={false} />
+        <BarPlot data={{ table: data }} units={selectedUnits} perCapita={false} />
       </div>
     </div>
   )
