+++ lib/StatePlots.js
@@ -3,12 +3,7 @@ import { useFetch } from '../lib/queries.js'
 import SelectSearch from 'react-select-search/dist/cjs'
 import { useMemo, useState } from 'react'
 import BarPlot from '../lib/BarPlot.js'
-
-const unitsOptions = [
-  { value: 'units', name: 'Units' },
-  { value: 'bldgs', name: 'Buildings' },
-  { value: 'value', name: 'Property value' }
-]
+import { makeUnitsSelect } from '../lib/selects.js'
 
 function makeStateOptions (statesList) {
   let stateNames = (
@@ -65,20 +60,6 @@ function makeStateSelect (stateOptions, stateName, router) {
   )
 }
 
-function makeUnitsSelect () {
-  const [selectedUnits, setSelectedUnits] = useState('units')
-  const unitsSelect = (
-    <SelectSearch
-      value={selectedUnits}
-      onChange={setSelectedUnits}
-      options={unitsOptions}
-    />
-  )
-  return {
-    selectedUnits, unitsSelect
-  }
-}
-
 export default function StatePlots ({ stateName }) {
   const router = useRouter()
 
@@ -105,9 +86,7 @@ export default function StatePlots ({ stateName }) {
 
       <div className='flex flex-col lg:grid lg:grid-cols-3'>
         <div className='m-4 col-span-1'>{stateSelect}</div>
-
         <h1 className='mt-4 text-4xl col-span-1 text-center'>{stateName}</h1>
-
         <div className='col-span-1 m-4'>{unitsSelect}</div>
       </div>
