+++ load_data.sh
@@ -1,4 +1,4 @@
 #! /usr/bin/env bash
-pushd python/housing-data/ || exit
+pushd python/ || exit
 poetry run build_data
 popd || exit
